add cmd mutation
