+++ exec/go.mod
@@ -1,3 +1,12 @@
 module github.com/defenseunicorns/pkg/exec
 
 go 1.21.8
+
+require github.com/defenseunicorns/pkg/helpers v1.1.1
+
+require (
+	github.com/otiai10/copy v1.14.0 // indirect
+	golang.org/x/sync v0.6.0 // indirect
+	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
+	oras.land/oras-go/v2 v2.5.0 // indirect
+)
